fix(map): refractored hash/equality tests and changed hash test
